Basic Graph Data Generation
Created Sage Program that can generate and save images of graphs with a given range of vertices.
Iterates through graphs with a number of edges equal to the number of vertices up to one less than the maximum number of edges.
Added folder with example graph data generated by this program.